requirements.txt
attrs
slackclient
